schema link for pre-renders during build
