+++ package.json
@@ -42,6 +42,7 @@
     "glamor": "^2.20.22",
     "graphql": "^14.7.0",
     "graphql-tag": "^2.11.0",
+    "graphql-tools": "^4.0.8",
     "gsheets": "^2.0.0",
     "isomorphic-unfetch": "^3.1.0",
     "lru-cache": "^5.1.1",
