+++ pages/404.js
@@ -3,7 +3,7 @@ import React from 'react'
 import Frame, {Center} from 'src/components/Frame'
 import MetaTags from 'src/components/MetaTags'
 import {H1, P} from 'src/components/Styled'
-import {createGetStaticProps} from 'lib/apolloClient'
+import {createGetStaticProps} from 'lib/apolloClientSchemaLink'
 
 export const NotFound = () => <>
   <MetaTags locale='de' title='404' />
